Proper Author Configuration

+++ src/main/java/com/nixon/myreads/client/BookResponse.java
@@ -4,6 +4,6 @@ import com.nixon.myreads.client.response.AuthorsResponse;
 
 import java.util.List;
 
-public record BookResponse(int id, String title, String image, List<AuthorsResponse> authors){
+public record BookResponse(Long id, String title, String image, List<AuthorsResponse> authors){
     
 }
